requirements.txt
pybaseball>=2.2.5
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
